Unify node/browser dispatcher code
Instead of having a mock Dispatcher for node, pull out the relevant code
into a small wrapper, web_worker.js, which contains the node/browser
code split to a smaller scope.
