+++ js/util/browser/web_worker.js
@@ -1,0 +1,6 @@
+'use strict';
+var WebWorkify = require('webworkify');
+
+module.exports = function () {
+    return new WebWorkify(require('../../source/worker'));
+};
