+++ js/util/dispatcher.js
@@ -1,57 +1,73 @@
 'use strict';
 
-var Worker = require('../source/worker');
-var Actor = require('../util/actor');
+var util = require('./util');
+var Actor = require('./actor');
+var WebWorker = require('./web_worker');
 
 module.exports = Dispatcher;
 
-function MessageBus(addListeners, postListeners) {
-    return {
-        addEventListener: function(event, callback) {
-            if (event === 'message') {
-                addListeners.push(callback);
-            }
-        },
-        postMessage: function(data) {
-            setImmediate(function() {
-                for (var i = 0; i < postListeners.length; i++) {
-                    postListeners[i]({data: data, target: this.target});
-                }
-            }.bind(this));
-        }
-    };
-}
-
+/**
+ * Responsible for sending messages from a {@link Source} to an associated
+ * {@link WorkerSource}.
+ *
+ * @interface Dispatcher
+ * @private
+ */
 function Dispatcher(length, parent) {
-
-    this.actors = new Array(length);
-
-    var parentListeners = [],
-        workerListeners = [],
-        parentBus = new MessageBus(workerListeners, parentListeners),
-        workerBus = new MessageBus(parentListeners, workerListeners);
-
-    parentBus.target = workerBus;
-    workerBus.target = parentBus;
-    // workerBus substitutes the WebWorker global `self`, and Worker uses
-    // self.importScripts for the 'load worker source' target.
-    workerBus.importScripts = function () {};
-
-    this.worker = new Worker(workerBus);
-    this.actor = new Actor(parentBus, parent);
-
-    this.remove = function() {
-        parentListeners.splice(0, parentListeners.length);
-        workerListeners.splice(0, workerListeners.length);
-    };
+    this.actors = [];
+    this.currentActor = 0;
+    for (var i = 0; i < length; i++) {
+        var worker = new WebWorker();
+        var actor = new Actor(worker, parent);
+        actor.name = "Worker " + i;
+        this.actors.push(actor);
+    }
 }
 
 Dispatcher.prototype = {
-    broadcast: function(type, data, callback) {
-        this.actor.send(type, data, callback);
+    /**
+     * Broadcast a message to all Workers.
+     * @method
+     * @name broadcast
+     * @param {string} type
+     * @param {object} data
+     * @param {Function} callback
+     * @memberof Dispatcher
+     * @instance
+     */
+    broadcast: function(type, data, cb) {
+        cb = cb || function () {};
+        util.asyncAll(this.actors, function (actor, done) {
+            actor.send(type, data, done);
+        }, cb);
     },
 
+    /**
+     * Send a message to a Worker.
+     * @method
+     * @name send
+     * @param {string} type
+     * @param {object} data
+     * @param {Function} callback
+     * @param {number|undefined} [targetID] The ID of the Worker to which to send this message. Omit to allow the dispatcher to choose.
+     * @returns {number} The ID of the worker to which the message was sent.
+     * @memberof Dispatcher
+     * @instance
+     */
     send: function(type, data, callback, targetID, buffers) {
-        this.actor.send(type, data, callback, buffers);
+        if (typeof targetID !== 'number' || isNaN(targetID)) {
+            // Use round robin to send requests to web workers.
+            targetID = this.currentActor = (this.currentActor + 1) % this.actors.length;
+        }
+
+        this.actors[targetID].send(type, data, callback, buffers);
+        return targetID;
+    },
+
+    remove: function() {
+        for (var i = 0; i < this.actors.length; i++) {
+            this.actors[i].target.terminate();
+        }
+        this.actors = [];
     }
 };
